requirements.txt
requests==2.22.0
matplotlib==3.4.3
python-decouple==3.5
pymongo==3.12.0
dnspython==2.1.0
